requirements.txt
pymodbus
pyserial
rich